feat: implemented monthly stock

+++ src/main/java/com/mercadolibro/dto/AuthorBookCountDTO.java
@@ -5,6 +5,5 @@ import io.swagger.annotations.ApiModel;
 @ApiModel(description = "Author total books count", value = "AuthorBookCountDTO")
 public interface AuthorBookCountDTO {
     String getAuthor();
-
-    Integer getTotal_books();
+    Long getTotal_books();
 }
