+++ src/main/java/com/mercadolibro/dto/PaymentTypeSaleDTO.java
@@ -5,7 +5,6 @@ import io.swagger.annotations.ApiModel;
 @ApiModel(description = "Total sales by payment type", value = "PaymentTypeSaleDTO")
 public interface PaymentTypeSaleDTO {
     PaymentMethod getPayment_type();
-
-    Integer getSales();
+    Long getSales();
 }
